requirements.txt
streamlit
python-dotenv
pymysql
pandas
apscheduler
matplotlib
seaborn